Calendar and Task updated

+++ src/pages/Task.js
@@ -17,30 +17,92 @@ import {
 import EditTaskModal from "../components/EditTaskModal.js";
 import classnames from 'classnames';
 import { FaPen, FaRegTrashAlt } from "react-icons/fa";
+import Calendar from 'react-calendar';
+import { toast } from 'react-toastify';
 
 
-const Task = () => {
+const Task = ({ data, setData }) => {
 
     const [tasks, setTasks] = useState([]);
 
     //  open modal boolean
     const [isOpen, setIsOpen] = useState(false);
 
-    //  
     const toggle = () => setIsOpen(!isOpen);
 
     const openEditForm = () => {
         setIsOpen(!isOpen);
     }
 
-    const testActivity = "Do laundry";
-    const testDate = "2021-02-05";
-
+    // Incomplete & Completed tabs
     const [activeTab, setActiveTab] = useState('1');
-
     const toggleTab = tab => {
         if (activeTab !== tab) setActiveTab(tab);
     }
+
+    const [dateValue, setDateValue] = useState(new Date());
+    const [titleValue, setTitleValue] = useState("");
+    // const [isCompleteValue, setIsCompleteValue] = useState(false);
+
+    // const toggleIsComplete = (e) => {
+    //     e.preventDefault();
+    //     setIsCompleteValue(e.target.value);
+    // }
+
+
+    const taskInputOnChange = (e) => {
+        e.preventDefault();
+        setTitleValue(e.target.value);
+    }
+
+    const addTaskOnSubmit = (e) => {
+        e.preventDefault();
+
+        // this is to get the total count of the data list 
+        let count = data.length;
+
+        // next id = count++
+        // create new data
+        let newData = {
+            id: count + 1,
+            title: titleValue,
+            start: dateValue,
+            end: dateValue,
+            isComplete: false,
+            allDay: true
+        }
+
+        if (titleValue != "") {
+            // push new data to list
+            setData([...data, newData]);
+
+            // reset the value on submit
+            setTitleValue("");
+            setDateValue(new Date());
+
+            toast.success(`New task added successfully`, {
+                position: "top-center",
+                autoClose: 5000,
+                hideProgressBar: false,
+                closeOnClick: true,
+                pauseOnHover: true,
+                draggable: true,
+                progress: undefined,
+            });
+        }
+        else {
+            toast.error(`Cannot add an empty task`, {
+                position: "top-center",
+                autoClose: 5000,
+                hideProgressBar: false,
+                closeOnClick: true,
+                pauseOnHover: true,
+                draggable: true,
+                progress: undefined,
+            });
+        }
+    }
+
 
     return (
         <div className="content">
@@ -49,16 +111,21 @@ const Task = () => {
                     <Row className="no-gutter">
                         <Col>
                             <h1 style={{ color: "#F0C38E", fontWeight: "bold" }}>Task</h1>
-                            <Form>
+                            <Form onSubmit={addTaskOnSubmit}>
                                 <FormGroup>
                                     <div className="task-input-form" style={{ marginBottom: "10px" }}>
                                         <Label for="task">Task:</Label>
-                                        <Input placeholder="Add a task" />
+                                        <Input type="text" placeholder="Add a task" value={titleValue} onChange={taskInputOnChange} />
                                     </div>
 
                                     <div className="date-input-form">
                                         <Label for="date">To be completed by:</Label>
-                                        <Input type="date" />
+                                        <p style={{ fontWeight: "bold", border: "solid 1px grey" }}>{dateValue.toLocaleDateString()}</p>
+                                        <Calendar
+                                            onChange={setDateValue}
+                                            value={dateValue}
+                                            minDate={new Date()}
+                                        />
                                     </div>
 
                                 </FormGroup>
@@ -90,33 +157,46 @@ const Task = () => {
                             <TabPane tabId="1">
                                 <Row>
                                     <Col>
-                                        <div className="task">
-                                            <div className="task-list"><Input type="checkbox" />Do laundry</div>
-                                            <div className="list-btn" style={{ paddingRight: "50px" }}>
-                                                <Button className="delete-btn" type="delete" color="danger"><FaRegTrashAlt /></Button>{' '}
-                                                <Button className="edit-btn" type="edit" color="primary" onClick={openEditForm}><FaPen /></Button>{' '}
-                                            </div>
-                                        </div>
-                                        <div className="task">
-                                            <div className="task-list"><Input type="checkbox" />Buy toiletries</div>
-                                            <div className="list-btn" style={{ paddingRight: "50px" }}>
-                                                <Button className="delete-btn" type="delete" color="danger"><FaRegTrashAlt /></Button>{' '}
-                                                <Button className="edit-btn" type="edit" color="primary" onClick={openEditForm}><FaPen /></Button>{' '}
-                                            </div>
-                                        </div>
+                                        {
+                                            data.map(x => {
+                                                if (!x.isComplete) {
+                                                    return (
+                                                        <div key={x.id} className="task">
+                                                            <div className="task-list"><Input type="checkbox" defaultChecked={x.isComplete} />{x.title}</div>
+                                                            <i>{x.startDate}</i>
+                                                            <div className="list-btn" style={{ paddingRight: "50px" }}>
+                                                                <Button className="delete-btn" type="delete" color="danger"><FaRegTrashAlt /></Button>{' '}
+                                                                <Button className="edit-btn" type="edit" color="primary" onClick={openEditForm}><FaPen /></Button>{' '}
+                                                            </div>
+                                                        </div>
+                                                    )
+                                                }
+                                                else { return (null) }
+                                            })
+                                        }
                                     </Col>
                                 </Row>
                             </TabPane>
                             <TabPane tabId="2">
                                 <Row>
                                     <Col>
-                                        <div className="task">
-                                            <div className="task-list"><Input type="checkbox" checked={true} />Feed the turtle</div>
-                                            <div className="list-btn" style={{ paddingRight: "50px" }}>
-                                                <Button className="delete-btn" type="delete" color="danger"><FaRegTrashAlt /></Button>{' '}
-                                                <Button className="edit-btn" type="edit" color="primary" onClick={openEditForm}><FaPen /></Button>{' '}
-                                            </div>
-                                        </div>
+                                        {
+                                            data.map(x => {
+                                                if (x.isComplete) {
+                                                    return (
+                                                        <div key={x.id} className="task">
+                                                            <div className="task-list"><Input type="checkbox" defaultChecked={x.isComplete} />{x.title}</div>
+                                                            <i>{x.startDate}</i>
+                                                            <div className="list-btn" style={{ paddingRight: "50px" }}>
+                                                                <Button className="delete-btn" type="delete" color="danger"><FaRegTrashAlt /></Button>{' '}
+                                                                <Button className="edit-btn" type="edit" color="primary" onClick={openEditForm}><FaPen /></Button>{' '}
+                                                            </div>
+                                                        </div>
+                                                    )
+                                                }
+                                                else { return (null) }
+                                            })
+                                        }
                                     </Col>
                                 </Row>
                             </TabPane>
@@ -126,8 +206,9 @@ const Task = () => {
                 <EditTaskModal
                     isOpen={isOpen}
                     toggle={toggle}
-                    taskInput={testActivity}
-                    dateInput={testDate}
+
+                // taskInput={testActivity}
+                // dateInput={testDate}
                 />
             </div>
         </div>
